Pizza nutrition and pricing data

Nutritional information and local pricing for core pizzas available at Little Caesars (Canada).

Usage

pizzas

Format

A tibble with 27 rows and 18 columns:

size

The size of the pizza. A size of NA means that the pizza only comes in one size.

crust

The type of crust used.

flavour

The name of the pizza.

price

Local pricing in Canadian Dollars.

calories

Total calories.

total_fat

Total fat, in grams.

sat_fat

Saturated fat, in grams.

trans_fat

Trans fat, in grams.

cholesterol

Cholesterol, in milligrams.

sodium

Sodium, in milligrams.

total_carb

Total carbohydrate, in grams.

fibre

Dietary fibre, in grams.

sugar

Sugar, in grams.

protein

Protein, in grams.

vit_a

Vitamin A, as a percentage of the recommended daily value.

vit_c

Vitamin C, as a percentage of the recommended daily value.

calcium

Calcium, as a percentage of the recommended daily value.

iron

Iron, as a percentage of the recommended daily value.

Source

https://littlecaesars.ca/en-ca/nutrition/

Details

Local pricing obtained in Ottawa, Canada (note: prices may vary within a city).